+++ webpack.config.js
@@ -4,7 +4,6 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 module.exports = {
   mode: 'production',
   entry: './src/index.js',
-  devtool: 'inline-source-map',
   devServer: {
     static: './dist',
   },
